Moved function into controller
