+++ routes/geminiConnection.js
@@ -1,45 +1,19 @@
 const express = require("express");
 const app = express();
 app.use(express.json());
-require("dotenv").config();
 const router = express.Router();
-
-// connect to gemini
-const { GoogleGenerativeAI } = require("@google/generative-ai");
-
-// API key
-const genAI = new GoogleGenerativeAI(process.env.API_KEY);
+const geminiConnection = require("../controllers/geminiConnectionFunc.js");
 
 router.post("/api/gemini-connection", async (req, res) => {
   const title = req.body.title;
   let message;
   req.body.message ? (message = req.body.message) : (message = "begin");
+  const history = req.body.history;
 
-  // The Gemini model
-  const model = genAI.getGenerativeModel({
-    model: "gemini-1.5-flash",
-    systemInstruction: `Turners Car Auctions is a market leader in the automotive industry in New Zealand, responsible for more than 10% of all used vehicle sales in the country.
-    Turners is currently undertaking a major re-design of its motor vehicle insurance systems and processes, including an investigation of ways in which the introduction of new technologies can enhance the buyer experience.
-    Turners Car Insurance is an internal division of Turners Car Auctions.
-    Due to restructuring a lot of staff will need to be re-trained into different rolls within the company.
-    You are Zane, an executive at Turners Car Insurance.
-    You are interviewing an employee for a ${title} role.
-    Start by asking, "Tell me about yourself." Then ask at least 6 questions relevant to the role. Adjust your questions based on the user's responses`,
+  // Use function from controller
+  geminiConnection(message, history, title).then((result) => {
+    res.send(result);
   });
-
-  let chat;
-
-  if (req.body.history) {
-    chat = model.startChat({ history: req.body.history });
-  } else {
-    chat = model.startChat();
-  }
-
-  const result = await chat.sendMessage(message);
-  const response = await result.response;
-  const text = response.text();
-
-  res.send(text);
 });
 
 module.exports = router;
